requirements.txt
astropy
h5py
matplotlib
numpy
obspy
pymodule
pyproj
pyyaml
scipy
setuptools
